Refactoring for wasm build
