+++ Cargo.toml
@@ -1,40 +1,9 @@
-[package]
-name = "ledger"
-description = "Library to exchange APDU with Ledger Nano S devices"
-version = "0.3.0"
-license = "Apache-2.0"
-authors = ["Zondax GmbH <info@zondax.ch>"]
-homepage = "https://github.com/zondax/ledger-rs"
-repository = "https://github.com/zondax/ledger-rs"
-readme = "README.md"
-categories  = ["authentication", "cryptography"]
-keywords = ["ledger", "nano", "blue", "apdu"]
-edition     = "2018"
-autobenches = false
+[workspace]
+members = [
+    "ledger-generic-rs",
+    "ledger-transport-rs"
+]
 
-[badges]
-circle-ci = { repository = "zondax/ledger-rs" }
-
-[lib]
-name = "ledger"
-
-[dependencies]
-byteorder = "1.2.4"
-libc = "0.2.50"
-thiserror = "1.0.10"
-cfg-if = "0.1.7"
-lazy_static = "1.3.0"
-hex = "0.3.2"
-log = "0.4"
-
-[target.'cfg(target_os = "linux")'.dependencies]
-nix = "0.13.0"
-
-[dependencies.hidapi]
-version = "1.1.1"
-default-features = false
-features=["linux-static-hidraw"]
-
-[dev-dependencies]
-serial_test = "0.3.2"
-env_logger = "0.7"
+[profile.release]
+# Tell `rustc` to optimize for small code size.
+opt-level = "s"
