+++ ledger-generic-rs/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "ledger-generic"
+version = "0.1.0"
+authors = ["Zondax GmbH <info@zondax.ch>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "ledger_generic"
