+++ ledger-generic-rs/README.md
@@ -1,0 +1,5 @@
+# ledger-generic-rs
+
+[![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
+
+Generic utilitary rust library used to communicate with Nano S/X devices
