+++ ledger-transport-rs/README.md
@@ -1,0 +1,5 @@
+# ledger-transport-rs
+
+[![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
+
+Communication library between Rust and Ledger Nano S/X devices
